Fix test and merge Controller
